Add layer-level unit test
