+++ megatron/core/transformer/experimental_attention_variant/dsa_fused_kernels/__init__.py
@@ -1,0 +1,13 @@
+from .indexer_bwd import indexer_bwd_interface
+from .indexer_topk_reducesum import indexer_topk_reducesum_interface
+from .sparse_mla_bwd import sparse_mla_bwd_interface
+from .sparse_mla_fwd import sparse_mla_fwd_interface
+from .sparse_mla_topk_reducesum import sparse_mla_topk_reducesum_interface
+
+__all__ = [
+    "indexer_bwd_interface",
+    "indexer_topk_reducesum_interface",
+    "sparse_mla_bwd_interface",
+    "sparse_mla_fwd_interface",
+    "sparse_mla_topk_reducesum_interface",
+]
